requirements.txt
Flask==3.1.1
flasgger==0.9.7.1
python-dotenv==1.1.0
Jinja2==3.1.6
